pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pipeprobe"
version = "0.1.0"
description = "Regression-aware LLM evaluation for AI systems built on data infrastructure"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "evaluation", "regression", "dbt", "airflow", "data-engineering", "ai-testing", "ci-cd"]
dependencies = [
    "anthropic>=0.25.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pytest>=8.0.0",
    "sqlglot>=23.0.0",
]

[project.optional-dependencies]
openai    = ["openai>=1.0.0"]
dbt       = ["dbt-core>=1.7.0"]
airflow   = ["apache-airflow>=2.8.0"]
snowflake = ["snowflake-connector-python>=3.0.0"]
all       = ["pipeprobe[openai,dbt,airflow,snowflake]"]
dev       = ["pytest-asyncio", "pytest-cov", "ruff", "mypy", "pre-commit"]

[project.scripts]
pipeprobe = "pipeprobe.cli:app"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
